Rename s3mapprovider to s3provider
