+++ confmap/provider/s3provider/go.mod
@@ -1,4 +1,4 @@
-module github.com/open-telemetry/opentelemetry-collector-contrib/confmap/provider/s3mapprovider
+module github.com/open-telemetry/opentelemetry-collector-contrib/confmap/provider/s3provider
 
 go 1.18
