+++ confmap/provider/s3provider/provider_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package s3mapprovider
+package s3provider
 
 import (
 	"bytes"
