+++ confmap/provider/s3provider/testdata/otel-config.yaml
